sched: walt: Fix the bug in initializing the new task demand
sysctl_sched_init_task_load_pct and sched_init_task_load_windows
are out of sync. Due to this the new task demand is not initialized
correctly. It is currently set to 0 due to which several new tasks
gets packed on the same CPU. This results in suboptimal performance
until load balancer sort it out.

Change-Id: Ibcc81ed7f3b532942fde8bf3bd148891d61a86bc
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Lau <laststandrighthere@gmail.com>