sched/fair: Skips the CPUs which can't fit the task
The current task placement code does not explicitly check
if a task fits on a CPU or not. Instead it relies on other
checks like new utilization is not above CPU's original
capacity and CPU has enough head room after accommodating
this task. These does not take the sched_upmigrate and
sched_downmigrate thresholds into account. Add an explicit
check to skip sched group that can't fit the task.

Change-Id: I49e9fdbe7490654d7b8fe08034dba2b29173e5db
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Lau <laststandrighthere@gmail.com>