sched/fair: Use CPU original capacity for max capacity CPU check
task_fits_max() checks if a task fits on a given CPU or not. Any task
can fit on the max capacity CPU. so task_fits_max() should return
true when called on a max capacity CPU. However the check inside
this funciton uses CPU capacity instead of original capacity when
comparing against max capacity. The CPU capacity can be lower than
the original capacity for a max capacity CPU due to RT tasks. In
such cases, we fail to upmigrate to the max capacity CPU.

Change-Id: I9f3c14b91d9a6d8a61bbd3a1fe7a8f4aa7415987
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Lau <laststandrighthere@gmail.com>