sched/fair: Avoid heavy packing of related thread groups tasks
find_best_target() selects an active CPU as the target CPU and
an idle CPU as the backup CPU. Which ever CPU saves the most
energy is finally selected. Currently we are skipping the
energy difference between the active and idle CPUs for tasks
belong to related thread groups. So we potentially miss out
an idle CPU if it saves more energy compared to an active
CPU. This happens when utilizing the active CPU keeps the
system at higher OPP.

Change-Id: Id0eef1c4bd01afda1cf36714fa1b2c3f257a8908
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Lau <laststandrighthere@gmail.com>