sched/fair: fix update_flags < RQCF_ACT_SKIP warning
A typical backtrace

 assert_clock_updated()
 rq_clock_task()
 task_hot()
 can_migrate_task()
 detach_tasks()
 load_balance()
 rebalance_domains()
 nohz_idle_balance()
 run_rebalance_domains()

WALT accounting introduces a double rq lock in detach_task. This
could lead to a release and reacquire of the rq lock. That release
could cause other context to grab it and clear the RQCF_UPDATED
flag.

So once the lock is reacquired check if the RQCF_UPDATED flag is lost
and if it is, call update_rq_clock() with sets the flag and updates
the rq clock too.

Change-Id: If829ddd23add02459da7876f42cd7dc55559ccc8
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Signed-off-by: Lau <laststandrighthere@gmail.com>