sched/fair: Consider an idle CPU outside c-state as an active CPU
The find_best_target() selects an active CPU and an idle CPU as
two candidate CPUs. Whichever CPUs saves the most energy compared
to the previous CPU is selected finally. An idle CPU i.e no runnable
tasks but also outside c-state is a good candidate to run the
waking task since the task can run immediately and there is no
idle exit latency. Hence consider such CPU as an active CPU which
helps both power and performance.

Change-Id: I34f40c2dbca70995a8e6b4a8d5876f802bc000bc
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
[satyap@codeaurora.org: Bring in is_packing_eligible functionality
from msm-4.9 to check whether packing can be done or not as part of
dependency]
Signed-off-by: Satya Durga Srinivasu Prabhala <satyap@codeaurora.org>
Signed-off-by: Lau <laststandrighthere@gmail.com>