ANDROID: sched/fair: prevent possible infinite loop in sched_group_energy
There is a race between hotplug and energy_diff which might result
in endless loop in sched_group_energy. When this happens, the end
condition cannot be detected.

We can store how many CPUs we need to visit at the beginning, and
bail out of the energy calculation if we visit more cpus than expected.

Bug: 72311797 72202633
Change-Id: I01c5a5ee9b9712903cb321d3b5ccf2539205d55d
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
Signed-off-by: Puja Gupta <pujag@codeaurora.org>
Signed-off-by: Lau <laststandrighthere@gmail.com>