sched: Define a stub function for sched_irqload when WALT is disabled
sched_irqload() is a WALT specific function which returns the average
IRQ load on a given CPU. It is accessed in sched_cpu_util trace point
for printing the IRQ load on different CPUs evaluated in CPU selection
algorithm. This trace point is not tied to WALT, so define a stub
function for sched_irqload().

Change-Id: Ie8f5365ddd5b7a4a0dea3a190b18c54d4d3c2d4b
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Lau <laststandrighthere@gmail.com>