sched/core: Fix an ordering issue in wake_up_new_task
If a task is activated before the window_start on the rq is initialized,
its ravg would get reset to 0. However, the init load on this task is
already pushed into the cumulative_runnable_avg. When the task is dequeued,
the task demand (which is reset to 0) will be removed from the
cumulative_runnable_avg, and as a result, the init load will be left
on the runqueue and cra will not be 0 even when CPU is idle.

Change-Id: I7619206dde275f7b64bfcfed206bf0c9a7245cf9
Signed-off-by: Runmin Wang <runminw@codeaurora.org>
Signed-off-by: Lau <laststandrighthere@gmail.com>