FIXUP: ANDROID: sched/fair: return idle CPU immediately for prefer_idle
When a CPU is selected for a prefer_idle task through find_best_target
this CPU will become a target CPU on which we perform an energy diff and
select between the previous CPU and the target CPU based on the estimated
energy consumption for both placements.

For prefer_idle tasks we should favour performance over energy savings
and therefore return a found idle CPU immediately.

Change-Id: I2242a51134ac172dd58b3b08375388a7b4b84af0
Signed-off-by: Ionela Voinescu <ionela.voinescu@arm.com>
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
Suggested-by: Joel Fernandes <joelaf@google.com>
Git-commit: dc6f7513623fe5fb4be5018cf334caee5b9ca039
Git-repo: http://android.googlesource.com/kernel/common
[adharmap@codeaurora.org: Continue to return prev_cpu instead of -1
in case we cannot find a suitable cpu to place the task. The variable
energy_cpu is kept around for just that.]
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Signed-off-by: Lau <laststandrighthere@gmail.com>