sched: restore all values for sched_{up,down}_migrate knobs in error case
Now that the sched_{up,down}_migrate knobs accept multiple values
depends on the number of clusters in the system, if any of the up
migrate values is more than respective down migrate value, updating
sched_{up,down}_migrate knobs are aborted by restoring only first
value and returning an error to user. Instead of just restoring first
value in the vector, restore all the values in case of error.

Change-Id: If565fb951aa0c00d2dc4a8b4f1f7a90caf4d325f
Signed-off-by: Satya Durga Srinivasu Prabhala <satyap@codeaurora.org>
Signed-off-by: Lau <laststandrighthere@gmail.com>